Deprecate & remove SetFile
