--- model/src/main/java/tcgone/carddb/model/SetFile.java
@@ -1,17 +1,0 @@
-package tcgone.carddb.model;
-
-import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-import java.util.List;
-
-public class SetFile {
-  public String filename;
-  public Set set;
-  public List<Card> cards;
-
-  public String toString() {
-    return ReflectionToStringBuilder.toString(this, ToStringStyle.MULTI_LINE_STYLE);
-  }
-
-}
