+++ tools/pom.xml
@@ -2,35 +2,38 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-<!--    <parent>-->
-<!--        <artifactId>carddb</artifactId>-->
-<!--        <groupId>net.tcgone</groupId>-->
-<!--        <version>2.0-SNAPSHOT</version>-->
-<!--    </parent>-->
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>net.tcgone.carddb</groupId>
+    <parent>
+        <groupId>net.tcgone.carddb</groupId>
+        <artifactId>carddb</artifactId>
+        <version>2.0-SNAPSHOT</version>
+    </parent>
     <artifactId>tools</artifactId>
-    <version>2.0-SNAPSHOT</version>
-
     <name>carddb-tools</name>
     <description>Toolset to deal with TCG ONE Card Repository</description>
 
-    <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.2</version>
-        <relativePath /> <!-- lookup parent from repository -->
-    </parent>
-
-    <properties>
-        <jackson.version>2.12.1</jackson.version>
-    </properties>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <!-- Import dependency management from Spring Boot -->
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-dependencies</artifactId>
+                <version>2.4.2</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>model</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>data</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
@@ -50,12 +53,7 @@
         <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
-            <version>${jackson.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-core</artifactId>
-            <version>${jackson.version}</version>
+            <version>${jackson-dataformats.version}</version>
         </dependency>
 
     </dependencies>
